regedit: dialog rework
This patch adds a new API for building dialogs. This helps
make new dialogs easier to create, because it provides some
common building blocks and a consistent way to deal with user
input.

Dialogs are divided into sections that stack vertically, and
common sections for typical UI things like text boxes, option lists,
and buttons are provided.

The old dialogs are rewritten to use this API.